refactor: update Predict consumers to import from types/
Switch controller and hook imports from providers/types to types/ for
externally-facing types. Remove Omit<> wrappers from PredictController
since controller-facing param types are now defined without signer.
