+++ app/components/UI/Predict/hooks/usePredictPlaceOrder.test.ts
@@ -3,8 +3,7 @@ import { useContext } from 'react';
 import { IconName } from '../../../../component-library/components/Icons/Icon';
 import { ToastVariants } from '../../../../component-library/components/Toast';
 import { DevLogger } from '../../../../core/SDKConnect/utils/DevLogger';
-import type { OrderPreview } from '../providers/types';
-import { Result, Side } from '../types';
+import { type OrderPreview, Result, Side } from '../types';
 import { usePredictPlaceOrder } from './usePredictPlaceOrder';
 import { usePredictTrading } from './usePredictTrading';
 import { usePredictBalance } from './usePredictBalance';
